refactor: added icon passed as props
